backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
httpx
google-genai
PyPDF2
Pillow
google-search-results
mcp
pandas
PyMuPDF
